+++ src/hklpy2/tests/test_misc.py
@@ -31,20 +31,20 @@ from ..devices import make_dynamic_instance
 from ..devices import parse_factory_axes
 from ..exceptions import NoForwardSolutions
 from ..exceptions import SolverError
-from ..misc import ConfigurationRunWrapper
 from ..misc import axes_to_dict
 from ..misc import compare_float_dicts
 from ..misc import convert_units
 from ..misc import distance_between_pos_tuples
 from ..misc import flatten_lists
-from ..misc import get_run_orientation
-from ..solver_utils import get_solver
 from ..misc import istype
-from ..misc import list_orientation_runs
 from ..misc import load_yaml_file
 from ..misc import pick_closest_solution
 from ..misc import pick_first_solution
 from ..misc import roundoff
+from ..run_utils import ConfigurationRunWrapper
+from ..run_utils import get_run_orientation
+from ..run_utils import list_orientation_runs
+from ..solver_utils import get_solver
 from ..tests.common import HKLPY2_DIR
 from ..typing import AnyAxesType
 from ..typing import AxesArray
